+++ pyproject.toml
@@ -1,40 +1,11 @@
 [project]
 name = "pii-detector"
 version = "0.1.0"
-description = "PII detection examples using Regex, Presidio, and spaCy NER"
+description = "PII detection"
 readme = "README.md"
-requires-python = ">=3.8"
+requires-python = ">=3.12"
 dependencies = [
+    "pyinstaller>=5.13.0",
     "spacy>=3.4.0",
-    "presidio-analyzer>=2.2.0",
-    "presidio-anonymizer>=2.2.0",
-    "pip>=25.0.1",
-    "pyinstaller>=6.0.0",
+    "spacy-curated-transformers",
 ]
-
-[project.optional-dependencies]
-dev = [
-    "pytest>=7.0.0",
-    "black>=22.0.0",
-    "ruff>=0.1.0",
-]
-advanced = [
-    "presidio-image-redactor>=0.0.50",
-    "torch>=1.9.0",
-    "transformers>=4.0.0",
-]
-
-[build-system]
-requires = ["hatchling"]
-build-backend = "hatchling.build"
-
-[tool.hatch.build.targets.wheel]
-packages = ["."]
-
-[tool.ruff]
-line-length = 88
-target-version = "py38"
-
-[tool.black]
-line-length = 88
-target-version = ['py38']
